Moved connection-behaviours into microbit-interfacing
