+++ src/__tests__/mocks/SpyConnectionBehaviour.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: MIT
  */
 
-import type ConnectionBehaviour from '../../script/connection-behaviours/ConnectionBehaviour';
+import type ConnectionBehaviour from '../../script/microbit-interfacing/connection-behaviours/ConnectionBehaviour';
 import MicrobitBluetooth from '../../script/microbit-interfacing/MicrobitBluetooth';
 import MBSpecs from '../../script/microbit-interfacing/MBSpecs';
