+++ src/script/microbit-interfacing/connection-behaviours/ConnectionBehaviours.ts
